% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{clean_cpm_df}
\alias{clean_cpm_df}
\title{this file will summarize the cpm data frame by separating out the tRNA reference name and summing the reads together where 'uncharged' was included in the refereence file name.}
\usage{
clean_cpm_df(cpm_df)
}
\arguments{
\item{cpm_df}{Counts per million dataframe from import_df()}
}
\value{
clean tRNA reference names and summarize redundant names
}
\description{
this file will summarize the cpm data frame by separating out the tRNA reference name and summing the reads together where 'uncharged' was included in the refereence file name.
}
