% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq-matricies.R
\name{count_matrix}
\alias{count_matrix}
\title{this file runs DESEQ2 analysis}
\usage{
count_matrix(clean_cpm_df, organism, metric)
}
\arguments{
\item{clean_cpm_df}{count per million dataframe with cleaned reference names}

\item{organism}{either host or phage. filters relevant colummns to generate single organism dataframe}

\item{metric}{either total_counts (abunance) or pct_charged (charging) (or just charged reads?)}
}
\value{
generate count matrix from clean_cpm_df
}
\description{
this file runs DESEQ2 analysis
}
