% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{import_df}
\alias{import_df}
\title{import data from files}
\usage{
import_df(sample_ids, replicates, filter_files_by)
}
\value{
A data frame containing sequencing metrics
}
\description{
this function will return one data frame from aa-tRNA-seq-pipeline
}
